# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Grow Flow AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /content/generate:
    post:
      operationId: generateContent
      tags: [content]
      summary: Generate multi-platform content
      description: Generate content for Instagram, YouTube Shorts, Twitter thread, and LinkedIn
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateContentBody"
      responses:
        "200":
          description: Generated content for all platforms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedContent"
        "401":
          description: Unauthorized
        "400":
          description: Bad request
...
  /improve-competitor:
    post:
      operationId: improveCompetitorContent
      tags: [content]
      summary: Improve competitor content
      description: Generate a better version of competitor content
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImproveCompetitorContentBody"
      responses:
        "200":
          description: Successful improvement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImproveCompetitorContentResult"
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    GenerateContentBody:
      type: object
      properties:
        idea:
          type: string
          description: The content idea to generate from
        contentType:
          type: string
          enum: [Educational, Story, Viral]
        tone:
          type: string
          enum: [Casual, Professional, Aggressive]
        language:
          type: string
          enum: [English, Hindi, Hinglish, Bengali, Gujarati, Kannada, Malayalam, Marathi, Punjabi, Odia, Assamese, Tamil, Telugu, Urdu, Bhojpuri, Spanish, French, German]
          description: Optional - language of the generated content
      required:
        - idea
        - contentType
        - tone
...
    PlatformContent:
      type: object
      properties:
        instagram:
          type: object
          properties:
            caption:
              type: string
            hashtags:
              type: string
          required:
            - caption
            - hashtags
        youtube:
          type: object
          properties:
            hook:
              type: string
            script:
              type: string
          required:
            - hook
            - script
        twitter:
          type: object
          properties:
            tweets:
              type: array
              items:
                type: string
          required:
            - tweets
        linkedin:
          type: object
          properties:
            post:
              type: string
          required:
            - post
      required:
        - instagram
        - youtube
        - twitter
        - linkedin
    GeneratedContent:
      type: object
      properties:
        id:
          type: integer
        idea:
          type: string
        contentType:
          type: string
        tone:
          type: string
        content:
          $ref: "#/components/schemas/PlatformContent"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - idea
        - contentType
        - tone
        - content
        - createdAt
...
    ImproveCompetitorContentBody:
      type: object
      properties:
        competitorContent:
          type: string
        language:
          type: string
          enum: [English, Hindi, Hinglish, Bengali, Gujarati, Kannada, Malayalam, Marathi, Punjabi, Odia, Assamese, Tamil, Telugu, Urdu, Bhojpuri, Spanish, French, German]
          description: Optional - language of the improved content
      required:
        - competitorContent
    ImproveCompetitorContentResult:
      type: object
      properties:
        competitorWeaknesses:
          type: array
          items:
            type: string
        missedAngle:
          type: string
        improvedVersion:
          type: string
        strongerHook:
          type: string
        newAngle:
          type: string
        monetization:
          type: object
          properties:
            ctaIdeas:
              type: array
              items:
                type: string
            affiliateSuggestions:
              type: array
              items:
                type: string
            howToEarn:
              type: array
              items:
                type: string
          required:
            - ctaIdeas
            - affiliateSuggestions
            - howToEarn
      required:
        - competitorWeaknesses
        - missedAngle
        - improvedVersion
        - strongerHook
        - newAngle
        - monetization