# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Api
 * Grow Flow AI API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * Generate content for Instagram, YouTube Shorts, Twitter thread, and LinkedIn
 * @summary Generate multi-platform content
 */
export const GenerateContentBody = zod.object({
  "idea": zod.string().describe('The content idea to generate from'),
  "contentType": zod.enum(['Educational', 'Story', 'Viral']),
  "tone": zod.enum(['Casual', 'Professional', 'Aggressive']),
  "language": zod.enum(['English', 'Hindi', 'Hinglish', 'Bengali', 'Gujarati', 'Kannada', 'Malayalam', 'Marathi', 'Punjabi', 'Odia', 'Assamese', 'Tamil', 'Telugu', 'Urdu', 'Bhojpuri', 'Spanish', 'French', 'German']).optional().describe('Optional - language of the generated content')
})

export const GenerateContentResponse = zod.object({
  "id": zod.number(),
  "idea": zod.string(),
  "contentType": zod.string(),
  "tone": zod.string(),
  "content": zod.object({
  "instagram": zod.object({
  "caption": zod.string(),
  "hashtags": zod.string()
}),
  "youtube": zod.object({
  "hook": zod.string(),
  "script": zod.string()
}),
  "twitter": zod.object({
  "tweets": zod.array(zod.string())
}),
  "linkedin": zod.object({
  "post": zod.string()
})
}),
  "createdAt": zod.coerce.date()
})


/**
 * Generate alternative versions of previously generated content
 * @summary Generate alternative content variations
 */
export const GenerateVariationsBody = zod.object({
  "historyId": zod.number().optional().describe('ID of the history item to generate variations for'),
  "idea": zod.string(),
  "contentType": zod.enum(['Educational', 'Story', 'Viral']),
  "tone": zod.enum(['Casual', 'Professional', 'Aggressive']),
  "language": zod.enum(['English', 'Hindi', 'Hinglish', 'Bengali', 'Gujarati', 'Kannada', 'Malayalam', 'Marathi', 'Punjabi', 'Odia', 'Assamese', 'Tamil', 'Telugu', 'Urdu', 'Bhojpuri', 'Spanish', 'French', 'German']).optional().describe('Optional - language of the generated content'),
  "platform": zod.enum(['instagram', 'youtube', 'twitter', 'linkedin']).optional().describe('Optional - specific platform to vary')
})

export const GenerateVariationsResponse = zod.object({
  "id": zod.number(),
  "idea": zod.string(),
  "contentType": zod.string(),
  "tone": zod.string(),
  "content": zod.object({
  "instagram": zod.object({
  "caption": zod.string(),
  "hashtags": zod.string()
}),
  "youtube": zod.object({
  "hook": zod.string(),
  "script": zod.string()
}),
  "twitter": zod.object({
  "tweets": zod.array(zod.string())
}),
  "linkedin": zod.object({
  "post": zod.string()
})
}),
  "createdAt": zod.coerce.date()
})


/**
 * Returns all past content generations for the authenticated user
 * @summary Get user's generation history
 */
export const getContentHistoryQueryLimitDefault = 20;
export const getContentHistoryQueryOffsetDefault = 0;

export const GetContentHistoryQueryParams = zod.object({
  "limit": zod.coerce.number().default(getContentHistoryQueryLimitDefault),
  "offset": zod.coerce.number().default(getContentHistoryQueryOffsetDefault)
})

export const GetContentHistoryResponse = zod.object({
  "items": zod.array(zod.object({
  "id": zod.number(),
  "idea": zod.string(),
  "contentType": zod.string(),
  "tone": zod.string(),
  "content": zod.object({
  "instagram": zod.object({
  "caption": zod.string(),
  "hashtags": zod.string()
}),
  "youtube": zod.object({
  "hook": zod.string(),
  "script": zod.string()
}),
  "twitter": zod.object({
  "tweets": zod.array(zod.string())
}),
  "linkedin": zod.object({
  "post": zod.string()
})
}),
  "createdAt": zod.coerce.date()
})),
  "total": zod.number()
})


/**
 * @summary Get a specific history item
 */
export const GetHistoryItemParams = zod.object({
  "id": zod.coerce.number()
})

export const GetHistoryItemResponse = zod.object({
  "id": zod.number(),
  "idea": zod.string(),
  "contentType": zod.string(),
  "tone": zod.string(),
  "content": zod.object({
  "instagram": zod.object({
  "caption": zod.string(),
  "hashtags": zod.string()
}),
  "youtube": zod.object({
  "hook": zod.string(),
  "script": zod.string()
}),
  "twitter": zod.object({
  "tweets": zod.array(zod.string())
}),
  "linkedin": zod.object({
  "post": zod.string()
})
}),
  "createdAt": zod.coerce.date()
})


/**
 * @summary Delete a history item
 */
export const DeleteHistoryItemParams = zod.object({
  "id": zod.coerce.number()
})

export const DeleteHistoryItemResponse = zod.object({
  "success": zod.boolean()
})


/**
 * Generate multiple viral hook options for a given topic
 * @summary Generate viral hooks
 */
export const GenerateHooksBody = zod.object({
  "topic": zod.string().describe('The topic to generate hooks for'),
  "tone": zod.enum(['Casual', 'Professional', 'Aggressive']),
  "language": zod.enum(['English', 'Hindi', 'Hinglish', 'Bengali', 'Gujarati', 'Kannada', 'Malayalam', 'Marathi', 'Punjabi', 'Odia', 'Assamese', 'Tamil', 'Telugu', 'Urdu', 'Bhojpuri', 'Spanish', 'French', 'German']).optional().describe('Optional - language of the generated hooks')
})

export const GenerateHooksResponse = zod.object({
  "hooks": zod.array(zod.string())
})


/**
 * Generate a better version of competitor content
 * @summary Improve competitor content
 */
export const ImproveCompetitorContentBody = zod.object({
  "competitorContent": zod.string(),
  "language": zod.enum(['English', 'Hindi', 'Hinglish', 'Bengali', 'Gujarati', 'Kannada', 'Malayalam', 'Marathi', 'Punjabi', 'Odia', 'Assamese', 'Tamil', 'Telugu', 'Urdu', 'Bhojpuri', 'Spanish', 'French', 'German']).optional().describe('Optional - language of the improved content')
})

export const ImproveCompetitorContentResponse = zod.object({
  "competitorWeaknesses": zod.array(zod.string()),
  "missedAngle": zod.string(),
  "improvedVersion": zod.string(),
  "strongerHook": zod.string(),
  "newAngle": zod.string(),
  "monetization": zod.object({
  "ctaIdeas": zod.array(zod.string()),
  "affiliateSuggestions": zod.array(zod.string()),
  "howToEarn": zod.array(zod.string())
})
})


/**
 * @summary Get content generation stats for the user
 */
export const GetContentStatsResponse = zod.object({
  "totalGenerations": zod.number(),
  "thisWeek": zod.number(),
  "topContentType": zod.string(),
  "topTone": zod.string(),
})